Update glfw3 to version 3.3
Version 3.3 of glfw adds support for querying the window content scale,
which glfw defines as "the ratio between the current DPI and the
platform's default DPI".  This is used to fix the UI elements' behavior
in macOS, which breaks because the window logical size usually differs
from OpenGL's framebuffer size by the window content scale (in HighDPI
screens).